Move 1.15.2 NMS
