--- Bukkit-1.15/pom.xml
@@ -1,128 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>Hyperverse</artifactId>
-        <groupId>se.hyperver.hyperverse</groupId>
-        <version>0.9.0-SNAPSHOT</version>
-        <relativePath>..</relativePath>
-    </parent>
-
-    <modelVersion>4.0.0</modelVersion>
-    <name>Bukkit 1.15</name>
-    <licenses>
-        <license>
-            <name>GNU General Public License Version 3</name>
-            <url>https://raw.githubusercontent.com/Sauilitired/Hyperverse/master/LICENSE</url>
-        </license>
-    </licenses><artifactId>Bukkit-1.15</artifactId>
-
-    <build>
-        <defaultGoal>clean package</defaultGoal>
-        <finalName>Hyperverse-${project.artifactId}-${project.version}</finalName>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.7.0</version>
-                <configuration>
-                    <source>${java.version}</source>
-                    <target>${java.version}</target>
-                    <compilerArgs>
-                        <arg>-parameters</arg>
-                    </compilerArgs>
-                    <annotationProcessorPaths>
-                        <path>
-                            <groupId>org.projectlombok</groupId>
-                            <artifactId>lombok</artifactId>
-                            <version>1.18.12</version>
-                        </path>
-                    </annotationProcessorPaths>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-shade-plugin</artifactId>
-                <version>3.2.4</version>
-                <executions>
-                    <execution>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>shade</goal>
-                        </goals>
-                        <configuration>
-                            <createDependencyReducedPom>false</createDependencyReducedPom>
-                        </configuration>
-                    </execution>
-                </executions>
-                <configuration>
-                    <dependencyReducedPomLocation>${project.build.directory}/dependency-reduced-pom.xml</dependencyReducedPomLocation>
-                    <relocations>
-                        <relocation>
-                            <pattern>io.papermc.lib</pattern>
-                            <shadedPattern>se.hyperver.hyperverse.paperlib</shadedPattern> <!-- Replace this -->
-                        </relocation>
-                        <relocation>
-                            <pattern>org.bstats</pattern>
-                            <shadedPattern>se.hyperver.hyperverse.bstats</shadedPattern>
-                        </relocation>
-                        <relocation>
-                            <pattern>co.aikar.commands</pattern>
-                            <shadedPattern>se.hyperver.hyperverse.aikar.commands</shadedPattern>
-                        </relocation>
-                        <relocation>
-                            <pattern>co.aikar.locales</pattern>
-                            <shadedPattern>se.hyperver.hyperverse.aikar.locales</shadedPattern>
-                        </relocation>
-                        <relocation>
-                            <pattern>co.aikar.util</pattern>
-                            <shadedPattern>se.hyperver.hyperverse.aikar.util</shadedPattern>
-                        </relocation>
-                        <relocation>
-                            <pattern>net.jodah.expiringmap</pattern>
-                            <shadedPattern>se.hyperver.hyperverse.expiringmap</shadedPattern>
-                        </relocation>
-                        <relocation>
-                            <pattern>me.minidigger.minimessage</pattern>
-                            <shadedPattern>se.hyperver.hyperverse.messages</shadedPattern>
-                        </relocation>
-                        <relocation>
-                            <pattern>net.kyori</pattern>
-                            <shadedPattern>se.hyperver.hyperverse.kyori</shadedPattern>
-                        </relocation>
-                        <relocation>
-                            <pattern>cloud.commandframework</pattern>
-                            <shadedPattern>se.hyperver.hyperverse.cloud</shadedPattern>
-                        </relocation>
-                        <relocation>
-                            <pattern>ninja.leaping.configurate</pattern>
-                            <shadedPattern>se.hyperver.hyperverse.configurate</shadedPattern>
-                        </relocation>
-                    </relocations>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-
-    <dependencies>
-        <dependency>
-            <groupId>se.hyperver.hyperverse</groupId>
-            <artifactId>Core</artifactId>
-            <version>0.9.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.spigotmc</groupId>
-            <artifactId>spigot</artifactId>
-            <version>1.15.2-R0.1-SNAPSHOT</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.jetbrains</groupId>
-            <artifactId>annotations</artifactId>
-            <version>20.1.0</version>
-            <scope>compile</scope>
-        </dependency>
-    </dependencies>
-
-</project>
