--- Bukkit-1.15/src/license/THIRD-PARTY.properties
@@ -1,26 +1,0 @@
-# Generated by org.codehaus.mojo.license.AddThirdPartyMojo
-#-------------------------------------------------------------------------------
-# Already used licenses in project :
-# - Apache License, Version 2.0
-# - GNU General Public License Version 3
-# - GNU General Public License, version 2 (GPL2), with the classpath exception
-# - ISC License
-# - MIT
-# - Public Domain
-# - The Apache Software License, Version 2.0
-# - The GNU Lesser General Public License v3.0
-# - The MIT License
-#-------------------------------------------------------------------------------
-# Please fill the missing licenses for dependencies :
-#
-#
-#Thu Apr 23 23:01:13 CEST 2020
-co.aikar--taskchain-bukkit--3.7.2=MIT
-co.aikar--DelegatingMap--1.0.0-20180331.052245-5=MIT
-co.aikar--acf-bukkit--0.5.0-20200418.224218-165=MIT
-co.aikar--locales--1.0-20181221.115311-17=MIT
-co.aikar--taskchain-core--3.7.2=MIT
-co.aikar--minecraft-timings--1.0.4=MIT
-co.aikar--Table--1.0.0-20180331.054128-7=MIT
-org.spigotmc--spigot--1.15.2-R0.1-SNAPSHOT=GNU General Public License v3.0
-co.aikar--acf-core--0.5.0-20200418.224207-160=MIT
