+++ Core/src/main/java/se/hyperver/hyperverse/modules/HyperverseModule.java
@@ -22,7 +22,9 @@ import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
 import com.google.inject.assistedinject.FactoryModuleBuilder;
+import javax.inject.Named;
 import org.bukkit.WorldCreator;
+import org.bukkit.plugin.Plugin;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import se.hyperver.hyperverse.util.HyperConfigShouldGroupProfiles;
 import se.hyperver.hyperverse.Hyperverse;
