+++ pom.xml
@@ -42,11 +42,11 @@
         <module>Core</module>
         <module>Bukkit-1.16.3</module>
         <module>Bukkit-1.16</module>
-        <module>Bukkit-1.15</module>
         <module>AnnotationProcessor</module>
 
         <module>NmsCommon</module>
         <module>Nms-1_14_4</module>
+        <module>Nms-1_15_2</module>
     </modules>
 
     <build>
